test: Update suite 5 and suite 3
Refactored test in suite five to account for incorrect practice of
throwing an error as a successful test. It splits the first test into
three tests total to test the yet to be implemented features of
adding a new feed, and rejecting feeds that do not have valid names
or urls. Two tests were added to suite three to test that the entries
loaded by loadFeed() are unique entries, and that the callback function is executed upon completion of the asynchronus function.